Moved services to shared folder, renamed ClientConnectionService to ConnectionService
