+++ .gitignore
@@ -27,3 +27,6 @@ npm-debug.log.*
 *.css.d.ts
 *.sass.d.ts
 *.scss.d.ts
+
+# Extra VS Code workspaces
+*.code-workspace
