+++ src/shared/services/ConnectionService.ts
