+++ src/shared/services/papi.ts
@@ -2,6 +2,6 @@
  * Unified module for accessing API features in extensions
  */
 
-import * as CommandService from '@services/CommandService';
+import * as CommandService from '@shared/services/CommandService';
 
 export default { commands: CommandService };
